Show thumbnails in upgrade board
