+++ tower-defense/src/upgrade_board/mod.rs
@@ -1,0 +1,4 @@
+mod components;
+mod data_conversion;
+
+pub use components::UpgradeBoardModal;
